% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopower.R
\name{autopower}
\alias{autopower}
\title{Power Required by Speed}
\usage{
autopower(V, m, A, cdrag = 0.3, crolling = 0.015, pair = 1.2, g = 9.8)
}
\arguments{
\item{m}{vehicle mass (kg)}

\item{A}{area of front of vehicle (m2)}

\item{cdrag}{coefficient due to drag default=0.3}

\item{crolling}{coefficient due to rolling/friction default=0.015}

\item{pair}{(kg/m3) default =1.2}

\item{g}{acceleration due to gravity (m/s) default=9.8}

\item{v}{vehicle speed (m/2)}
}
\value{
power (W)
}
\description{
This function determines the power required to keep a vehicle moving at
a given speed
}
