% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computediversity.R
\name{computediversity}
\alias{computediversity}
\title{Describe diversity based on a list of species}
\usage{
computediversity(species)
}
\arguments{
\item{species}{list of species (names, or code)}
}
\value{
list with the following items
\describe{
\item{num}{ Number of distinct species}
\item{simpson}{Value of simpson diversity index}
\item{dominant}{Name of the most frequently occuring species}
}
}
\description{
Compute a species diversity index
}
\examples{
computediversity(c("butterfly","butterfly","mosquito","butterfly","ladybug","ladybug")))
}
\references{
http://www.tiem.utk.edu/~gross/bioed/bealsmodules/simpsonDI.html
}
