% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM262Examples-package.R
\docType{package}
\name{ESM262Examples-package}
\alias{ESM262Examples}
\alias{ESM262Examples-package}
\title{ESM262Examples: Examples for ESM 262}
\description{
Provides example R code for ESM 262 and serves as an example package
}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Naomi Tague
}

}
\keyword{internal}
