% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish catch data}
\format{
A dataframe with 100 rows 1 column
\itemize{
\item fish name
}
}
\source{
\url{http://www.sblter.ed/}
}
\usage{
fish
}
\description{
Data on a catch of fish from Santa Barbara
}
\keyword{datasets}
