% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_gen_orig.R
\name{power_gen_orig}
\alias{power_gen_orig}
\title{Power Generated by Reservoir}
\usage{
power_gen_orig(height, flow, rho = 1000, g = 9.8, Keff = 0.8)
}
\arguments{
\item{height}{reservoir height (m)}

\item{flow}{rate (m/s)}

\item{rho}{density of water (kg/m3)}

\item{g}{acceleration due to gravity (m/s) default=9.8}

\item{Keff}{reservoir efficiency (0-1) default 0.8}
}
\value{
power (W)
}
\description{
This function determines the power generated by a given flow rate for a reservoir
}
