% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spring_summary.R
\name{spring_summary}
\alias{spring_summary}
\title{Summary information about spring climate}
\usage{
spring_summary(clim_data, spring_months = c(4:6), spring_out = FALSE)
}
\arguments{
\item{clim_data}{data frame with columns tavg (C)
rain (precip in mm), year, month (integer), day}

\item{spring_months}{(vector of integers) to include in spring; default 4,5,6}

\item{spring_out}{(default FALSE) set to TRUE to output spring precip and temperature for all years}
}
\value{
returns a list containing,
\describe{
\item{Tavg_spring}{mean_springT mean spring temperature (C)}
\item{Pavg_spring}{ mean spring precipitation  (mm)}
\item{Tmax_spring}{  highest spring temperature  (C)}
\item{Tmin_spring}{  lowest spring temperature  (C)}
\item{Pmax_spring}{highest spring precip (mm)}
\item{Pmin_spring}{ lowest spring precip (mm)}
\item{warmest_spring}{ year with highest spring temperature  (year)}
\item{coldest_spring}{ year with lowest spring temperature  (year)}
\item{wettest_spring}{ spring (as year) with highest precip (year)}
\item{driest_spring}{ spring (as year) with highest precip (year)}
and array of means for all years if spring_out set to TRUE
}
}
\description{
computes summary information about spring temperature and precipitation
}
